refactor and update did-login code
