+++ menus/exit-cli.js
@@ -1,0 +1,4 @@
+export default function () {
+  console.clear() 
+  process.exit() 
+}
