+++ recipes/nextjs-sporran-did-login/package.json
@@ -1,6 +1,5 @@
 {
   "name": "nextjs-did-login",
-  "private": true,
   "scripts": {
     "dev": "next dev",
     "build": "next build",
@@ -28,5 +27,8 @@
   "devDependencies": {
     "eslint": "8.5.0",
     "eslint-config-next": "12.0.7"
-  }
+  },
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT"
 }
