+++ recipes/nextjs-sporran-did-login/pages/api/logout.js
