+++ recipes/nextjs-sporran-did-login/pages/member/secret.js
