+++ recipes/nextjs-sporran-did-login/public/favicon.ico
